--- src/components/form/Textfield.tsx
@@ -1,16 +1,0 @@
-import { TextFieldProps as MUITextFieldProps } from '@mui/material'
-
-export enum TextFieldStatus {
-  error = 'error',
-  success = 'success',
-}
-
-interface TextFieldCSSProps {
-  status?: TextFieldStatus | undefined
-}
-
-type TextFieldProps = MUITextFieldProps & TextFieldCSSProps
-
-export default function TextField(props: TextFieldProps) {
-  return <TextField {...props} />
-}
